3 new weapons and 3 Hai ships
The new weapons were some things that I made during the whole gatling
gun scenario after Local asked if anyone could make a competing railgun.
The Mod Proton Gun was my first creation, and the two particle weapons
came afterwards when it was known that the created railgun would be
going to the Hai. I made two versions, close range and long range, just
because.

As for the Hai ships; they were created during the same timeframe by
Matthew as concepts for a Hai drone carrier. We eventually settled on a
design that Local created, but I though that Matthew's ships were good
enough to be put to some use. Images for them coming in the future after
I get the .blend and .xcf files to mess around with them a bit. :)