Complete Echo Series rebalance
For the most part, everything has been nerfed slightly so that they're
not all top of their class.

With the exception of the Oblivion and Aegis, they are mostly middle of
the pack to slightly above average.

Another change is that most of the Echo ships come stock with Navy
generators (Dwarf, Fission, Breeder, Fusion, Armageddon) and Ion
engines.