Small government changes
Removed ORM's affection of the Wanderers. No idea why I put that in
there. I've got no plans for ORM to come into contact with the Wanderers
anyway.
Also removed EC's hate of test dummies, so hopefully it doesn't
interfere with missions. That was just for one EC mission anyways, and I
fixed it by changing the related EC ships to the Escort government
instead.